Install typescript, jest and @types. Move js files to ts. Move jsx files to tsx.
