+++ package.json
@@ -16,6 +16,7 @@
     "expo": "~39.0.0",
     "expo-image-picker": "^9.1.1",
     "graphql": "^15.0.0",
+    "jest": "^26.6.3",
     "lodash.flowright": "^3.5.0",
     "moment": "^2.29.1",
     "react": "~16.13.0",
@@ -29,10 +30,16 @@
     "react-native-vector-icons": "^7.1.0",
     "react-native-web": "0.13.13",
     "subscriptions-transport-ws": "^0.9.18",
+    "typescript": "^4.1.3",
     "yup": "^0.29.3"
   },
   "devDependencies": {
     "@babel/core": "7.9.0",
+    "@types/graphql": "^14.5.0",
+    "@types/jest": "^26.0.19",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
+    "@types/react-navigation": "^3.4.0",
     "babel-eslint": "^10.1.0",
     "babel-plugin-import-graphql": "^2.7.0",
     "eslint": "^7.13.0",
