--- src/context.js
@@ -1,4 +1,0 @@
-import React from "react";
-
-export const MeContext = React.createContext();
-export const UserCheckoutNotificationContext = React.createContext(0);
