+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { ApolloProvider } from '@apollo/client';
-
+import { ApolloProvider as ApolloHooksProvider } from '@apollo/react-hoc'
 import { apolloClient } from '../apollo';
-import Navigation from '../src/navigation';
+import Navigation from './navigation';
 
 import { ThemeProvider } from 'react-native-elements';
 
@@ -11,9 +11,11 @@ import { reactNativeElementTheme } from './theme';
 export default () => {
   return (
     <ApolloProvider client={apolloClient}>
-      <ThemeProvider theme={reactNativeElementTheme}>
-        <Navigation />
-      </ThemeProvider>
+      <ApolloHooksProvider client={apolloClient as any}>
+        <ThemeProvider theme={reactNativeElementTheme}>
+          <Navigation />
+        </ThemeProvider>
+      </ApolloHooksProvider>
     </ApolloProvider>
   );
 }
