+++ src/orders/admin/allOrders.tsx
@@ -6,6 +6,7 @@ import { useQuery, gql } from '@apollo/client';
 import { colorsLocal } from '../../theme';
 import moment from "moment";
 import NEW_ORDER_SUBSCRIPTION from './latestOrder.graphql';
+import { NavigationScreenProp } from 'react-navigation';
 
 const GET_ORDERS_ADMIN_QUERY = gql`
   query {
@@ -33,7 +34,11 @@ const GET_ORDERS_ADMIN_QUERY = gql`
   } 
 `
 
-const AllOrders = ({ navigation }) => {
+interface Props {
+  navigation: NavigationScreenProp<any, any> | any;
+}
+
+const AllOrders: React.SFC<Props> = ({ navigation }) => {
   const { data, loading, error, subscribeToMore } = useQuery(GET_ORDERS_ADMIN_QUERY);
 
   if (error) {
@@ -92,14 +97,14 @@ const AllOrders = ({ navigation }) => {
         keyExtractor={(item) => item.id.toString()}
         ItemSeparatorComponent={renderSeprator}
         renderItem={({ item }) => (
-          <Card style={styles.card}>
+          <Card containerStyle={styles.card}>
             <Text style={styles.text}>
               <Text style={styles.label}>Name: </Text>{item.name}
             </Text>
             <Text>
               <Text style={styles.label}>Email: </Text>{item.email}
             </Text>
-            <Text style={styles.text, { marginBottom: 8 }}>
+            <Text style={styles.textPhone}>
               <Text style={styles.label}>Phone: </Text>{item.phone}
             </Text>
             <Card.Divider />
@@ -171,7 +176,7 @@ const AllOrders = ({ navigation }) => {
                   name="eye"
                   size={20}
                   style={{ marginRight: 10 }}
-                  color={colors.white}
+                  color='white'
                 />
               }
               onPress={() => { navigation.push('ViewUserOrdersAdmin', { name: 'View user orders (Admin view)', id: item.id }) }}
@@ -225,6 +230,10 @@ const styles = StyleSheet.create({
   },
   text: {
     textTransform: 'capitalize'
+  },
+  textPhone: {
+    textTransform: 'capitalize',
+    marginBottom: 8
   },
   currency: {
     fontSize: 12,
